requirements.txt
.idea
Django==1.11.2
djangorestframework==3.5.4
typing==3.5.3.0